Create Apple_and _Orange.py
Sam's house has an apple tree and an orange tree that yield an abundance of fruit. Using the information given below, determine the number of apples and oranges that land on Sam's house.

In the diagram below:

The red region denotes the house, where S is the start point, and T is the endpoint. The apple tree is to the left of the house, and the orange tree is to its right.
Assume the trees are located on a single point, where the apple tree is at point , and the orange tree is at point .
When a fruit falls from its tree, it lands D units of distance from its tree of origin along the X-axis. *A negative value of D means the fruit fell D units to the tree's left, and a positive value of D means it falls D units to the tree's right. *

Apple and orange(2).png

Function Description

Complete the countApplesAndOranges function in the editor below. It should print the number of apples and oranges that land on Sam's house, each on a separate line.

countApplesAndOranges has the following parameter(s):

S: integer, starting point of Sam's house location.
T: integer, ending location of Sam's house location.
A: integer, location of the Apple tree.
B: integer, location of the Orange tree.
apples: integer array, distances at which each apple falls from the tree.
oranges: integer array, distances at which each orange falls from the tree.

Input Format

The first line contains two space-separated integers denoting the respective values of S and T.
The second line contains two space-separated integers denoting the respective values of A and B.
The third line contains two space-separated integers denoting the respective values of M and N.
The fourth line contains  space-separated integers denoting the respective distances that each apple falls from point .
The fifth line contains  space-separated integers denoting the respective distances that each orange falls from point .